requirements.txt
ctransformers
sentence-transformers
langchain
flask
pypdf
chromadb
langchain-google-genai
-e .
python-dotenv